python/requirements.txt
numpy==1.18.1
tqdm==4.43.0
Pillow==7.0.0
tensorflow==2.1.0
matplotlib==3.2.1